chore: refactor main function
Refactor the main function by moving some of its logic into separate modules for better organization and readability.
